NMS-9335: LoadingTopology from db result in NPE
Fixed concurrent exception.
changed clean topology layout

+++ opennms-model/src/main/java/org/opennms/netmgt/model/topology/SharedSegment.java
@@ -298,10 +298,10 @@ public class SharedSegment {
             return false;
         }
         for (BridgePort port: m_portsOnSegment) {
-        	if (port.getNode().getId() != nodeid) {
+        	if (port.getNode().getId().intValue() != nodeid.intValue()) {
         		continue;
         	}
-        	if (port.getBridgePort() != bridgeport) {
+        	if (port.getBridgePort().intValue() != bridgeport.intValue()) {
         		continue;
         	}
         	return true;
